improve decimal-long consistency for monetary data

+++ src/AlgoTradeForge.Domain/Engine/BacktestEngine.cs
@@ -33,8 +33,8 @@ public sealed class BacktestEngine(IBarMatcher barMatcher, IRiskEvaluator riskEv
         var orderIdCounter = 0L;
         var totalBarsDelivered = 0;
         var activeSlTpPositions = new List<ActiveSlTpPosition>();
-        var lastPrices = new Dictionary<string, decimal>();
-        var equityCurve = new List<decimal>();
+        var lastPrices = new Dictionary<string, long>();
+        var equityCurve = new List<long>();
 
         strategy.OnInit();
 
@@ -92,7 +92,7 @@ public sealed class BacktestEngine(IBarMatcher barMatcher, IRiskEvaluator riskEv
                 strategy.OnBarComplete(bar, subscription, orderContext);
                 AssignOrderIds(orderQueue, ref orderIdCounter, barTimestamp);
 
-                lastPrices[subscription.Asset.Name] = (decimal)bar.Close;
+                lastPrices[subscription.Asset.Name] = bar.Close;
                 cursors[s]++;
                 totalBarsDelivered++;
             }
@@ -244,7 +244,7 @@ public sealed class BacktestEngine(IBarMatcher barMatcher, IRiskEvaluator riskEv
     private sealed class ActiveSlTpPosition
     {
         public required Order OriginalOrder { get; init; }
-        public required decimal EntryPrice { get; init; }
+        public required long EntryPrice { get; init; }
         public decimal RemainingQuantity { get; set; }
         public int NextTpIndex { get; set; }
     }
@@ -276,7 +276,7 @@ internal sealed class BacktestOrderContext : IOrderContext
         _portfolio = portfolio;
     }
 
-    public decimal Cash => _portfolio.Cash;
+    public long Cash => _portfolio.Cash;
 
     public void BeginBar(int currentFillCount)
     {
