+++ src/AlgoTradeForge.Domain/Engine/BacktestOptions.cs
@@ -4,11 +4,11 @@ namespace AlgoTradeForge.Domain.Engine;
 
 public sealed record BacktestOptions
 {
-    public required decimal InitialCash { get; init; }
+    public required long InitialCash { get; init; }
     public required Asset Asset { get; init; }
     public required DateTimeOffset StartTime { get; init; }
     public required DateTimeOffset EndTime { get; init; }
-    public decimal CommissionPerTrade { get; init; } = 0m;
+    public long CommissionPerTrade { get; init; } = 0L;
     public long SlippageTicks { get; init; }
     public bool UseDetailedExecutionLogic { get; init; }
 }
